Feat: oauth.interface.ts 추가
