+++ types/oauth..interface.ts
@@ -1,0 +1,10 @@
+export type OAuthPlatform = "kakao" | "google" | "bsm";
+
+export interface Token {
+  token: string;
+  validity: string;
+}
+
+export interface Bsm {
+  authCode: string;
+}
